pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "general-browser-ai-agent"
version = "0.1.0"
description = "Production-minded MVP for a browser-based AI research agent."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "alembic>=1.13.1",
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "playwright>=1.48.0",
  "pydantic-settings>=2.5.2",
  "python-multipart>=0.0.9",
  "sqlmodel>=0.0.22",
  "trafilatura>=1.12.2",
  "uvicorn[standard]>=0.31.0",
  "youtube-transcript-api>=0.6.2",
  "langgraph>=1.1.0",
  "langchain-core>=1.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.9",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
